Allow null CFP dates
Since we were using a default value (0000-00-00) for the CFP dates
before this, it was really unpleasant to check if a date had been set.
Also, the condition in `isCurrentlyAcceptingProposals` was impossible
to trigger.

I recommend another migration that converts any existing conferences
with `0000-00-00` style dates to null.

Also, notice the cool L5 change column feature. Taylor told me about it
last night but it was semi-broken until a PR I submitted was merged in
this afternoon :) Works with SQLite and everything!